+++ packages/stg-evm-v2/devtools/config/mainnet/01/feelib-v1.usdc.config.ts
@@ -17,6 +17,7 @@ import {
     onDegen,
     onEth,
     onFlare,
+    onFlow,
     onFuse,
     onGlue,
     onGravity,
@@ -63,6 +64,7 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
     const degenFeeLibV1 = onDegen(contract)
     const ethFeeLibV1 = onEth(contract)
     const flareFeeLibV1 = onFlare(contract)
+    const flowFeeLibV1 = onFlow(contract)
     const fuseFeeLibV1 = onFuse(contract)
     const glueFeeLibV1 = onGlue(contract)
     const gravityFeeLibV1 = onGravity(contract)
@@ -134,6 +136,10 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
             },
             {
                 contract: flareFeeLibV1,
+                config: defaultNodeConfig,
+            },
+            {
+                contract: flowFeeLibV1,
                 config: defaultNodeConfig,
             },
             {