+++ packages/stg-evm-v2/devtools/config/mainnet/01/usdt-token.config.ts
@@ -7,7 +7,7 @@ import { EndpointId } from '@layerzerolabs/lz-definitions'
 import { OwnableNodeConfig } from '@layerzerolabs/ua-devtools'
 
 import { createGetAssetAddresses, getAssetNetworkConfig } from '../../../../ts-src/utils/util'
-import { onAbstract, onDegen, onFuse, onGlue, onHemi, onIslander, onPeaq, onPlume, onRootstock } from '../utils'
+import { onAbstract, onDegen, onFlow, onFuse, onGlue, onHemi, onIslander, onPeaq, onPlume, onRootstock } from '../utils'
 
 const fiatContract = { contractName: 'TetherTokenV2' }
 
@@ -20,6 +20,9 @@ assert(usdtAbstractAsset.address != null, `External USDT address not found for A
 
 const usdtDegenAsset = getAssetNetworkConfig(EndpointId.DEGEN_V2_MAINNET, TokenName.USDT)
 assert(usdtDegenAsset.address != null, `External USDT address not found for DEGEN`)
+
+const usdtFlowAsset = getAssetNetworkConfig(EndpointId.FLOW_V2_MAINNET, TokenName.USDT)
+assert(usdtFlowAsset.address != null, `External USDT address not found for FLOW`)
 
 const usdtFuseAsset = getAssetNetworkConfig(EndpointId.FUSE_V2_MAINNET, TokenName.USDT)
 assert(usdtFuseAsset.address != null, `External USDT address not found for FUSE`)
@@ -56,6 +59,10 @@ export default async (): Promise<OmniGraphHardhat<OwnableNodeConfig, unknown>> =
         onDegen({ contractName: 'TransparentUpgradeableProxy', address: usdtDegenAsset.address })
     )
 
+    const flowUSDTProxy = await contractFactory(
+        onFlow({ contractName: 'TransparentUpgradeableProxy', address: usdtFlowAsset.address })
+    )
+
     const fuseUSDTProxy = await contractFactory(
         onFuse({ contractName: 'TransparentUpgradeableProxy', address: usdtFuseAsset.address })
     )
@@ -83,6 +90,7 @@ export default async (): Promise<OmniGraphHardhat<OwnableNodeConfig, unknown>> =
     const peaqUSDT = onPeaq({ ...fiatContract, address: peaqUSDTProxy.contract.address })
     const abstractUSDT = onAbstract({ ...fiatContract, address: abstractUSDTProxy.contract.address })
     const degenUSDT = onDegen({ ...fiatContract, address: degenUSDTProxy.contract.address })
+    const flowUSDT = onFlow({ ...fiatContract, address: flowUSDTProxy.contract.address })
     const fuseUSDT = onFuse({ ...fiatContract, address: fuseUSDTProxy.contract.address })
     const glueUSDT = onGlue({ ...fiatContract, address: glueUSDTProxy.contract.address })
     const hemiUSDT = onHemi({ ...fiatContract, address: hemiUSDTProxy.contract.address })
@@ -96,6 +104,7 @@ export default async (): Promise<OmniGraphHardhat<OwnableNodeConfig, unknown>> =
     const peaqAssetAddresses = await getAssetAddresses(EndpointId.PEAQ_V2_MAINNET, usdtAssets)
     const abstractAssetAddresses = await getAssetAddresses(EndpointId.ABSTRACT_V2_MAINNET, usdtAssets)
     const degenAssetAddresses = await getAssetAddresses(EndpointId.DEGEN_V2_MAINNET, usdtAssets)
+    const flowAssetAddresses = await getAssetAddresses(EndpointId.FLOW_V2_MAINNET, usdtAssets)
     const fuseAssetAddresses = await getAssetAddresses(EndpointId.FUSE_V2_MAINNET, usdtAssets)
     const glueAssetAddresses = await getAssetAddresses(EndpointId.GLUE_V2_MAINNET, usdtAssets)
     const hemiAssetAddresses = await getAssetAddresses(EndpointId.HEMI_V2_MAINNET, usdtAssets)
@@ -121,6 +130,12 @@ export default async (): Promise<OmniGraphHardhat<OwnableNodeConfig, unknown>> =
                 contract: degenUSDT,
                 config: {
                     owner: degenAssetAddresses.USDT,
+                },
+            },
+            {
+                contract: flowUSDT,
+                config: {
+                    owner: flowAssetAddresses.USDT,
                 },
             },
             {